Rename News headline extraction to News Headline Extraction
First, the program will read the news file and remove the <br /> tag.Then, the program will divide the news into meaningful chinese words and put it into the worddict dictionary. If the word already in the dictionary, the value of it will be plus one. After that, the program will use the built-in sorting method to sort the chinese words according to their occurrences. Finally, the program will write the result to the text file.